+++ src/dependency/Dockerfile
@@ -24,9 +24,8 @@ ENV DEBUG= \
     DEPENDENCY_QUEUE_URL= \
     MAX_NUMBER_OF_MESSAGE= \
     WAIT_TIME_SECOND=20 \
-    FINDING_SVC_ADDR= \
-    ALERT_SVC_ADDR= \
-    CODE_SVC_ADDR= \
+    CORE_SVC_ADDR= \
+    DATA_SOURCE_API_SVC_ADDR= \
     GITHUB_DEFAULT_TOKEN= \
     DATA_KEY= \
     LIMIT_REPOSITORY_SIZE_KB= \
