+++ src/dependency/handler.go
@@ -16,6 +16,8 @@ import (
 	"github.com/ca-risken/datasource-api/pkg/message"
 	"github.com/ca-risken/datasource-api/proto/code"
 	"github.com/google/go-github/v44/github"
+	"google.golang.org/grpc"
+	"google.golang.org/grpc/credentials/insecure"
 )
 
 type sqsHandler struct {
@@ -26,6 +28,17 @@ type sqsHandler struct {
 	alertClient           alert.AlertServiceClient
 	codeClient            code.CodeServiceClient
 	limitRepositorySizeKb int
+}
+
+func getGRPCConn(ctx context.Context, addr string) (*grpc.ClientConn, error) {
+	// gRPCクライアントの呼び出し回数が非常に多くトレーシング情報の送信がエラーになるため、トレースは無効にしておく
+	ctx, cancel := context.WithTimeout(ctx, 3*time.Second)
+	defer cancel()
+	conn, err := grpc.DialContext(ctx, addr, grpc.WithTransportCredentials(insecure.NewCredentials()))
+	if err != nil {
+		return nil, err
+	}
+	return conn, nil
 }
 
 func newHandler(ctx context.Context, conf *AppConfig) *sqsHandler {
